fix HTTP 404 caused by burst polling on s3 endpoint
