+++ internal/server/handlers/send/send_service.go
@@ -1,7 +1,6 @@
 package send
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
@@ -12,8 +11,6 @@ import (
 	"path"
 	"time"
 
-	"github.com/openmined/syftbox/internal/server/blob"
-	"github.com/openmined/syftbox/internal/server/handlers/ws"
 	"github.com/openmined/syftbox/internal/syftmsg"
 	"github.com/openmined/syftbox/internal/utils"
 )
@@ -23,61 +20,6 @@ var (
 	ErrNoRequest   = errors.New("no request found")
 )
 
-type BlobMsgStore struct {
-	blob *blob.BlobService
-}
-
-func (m *BlobMsgStore) GetMsg(ctx context.Context, path string) (io.ReadCloser, error) {
-	object, err := m.blob.Backend().GetObject(ctx, path)
-	if err != nil {
-		return nil, err
-	}
-	return object.Body, nil
-}
-
-func (m *BlobMsgStore) DeleteMsg(ctx context.Context, path string) error {
-	_, err := m.blob.Backend().DeleteObject(ctx, path)
-	return err
-}
-
-func (m *BlobMsgStore) StoreMsg(ctx context.Context, path string, msg syftmsg.SyftRPCMessage) error {
-	msgBytes, err := msg.MarshalJSON()
-	if err != nil {
-		return fmt.Errorf("failed to marshal message: %w", err)
-	}
-
-	_, err = m.blob.Backend().PutObject(ctx, &blob.PutObjectParams{
-		Key:  path,
-		ETag: msg.ID.String(),
-		Body: bytes.NewReader(msgBytes),
-		Size: int64(len(msgBytes)),
-	})
-	return err
-}
-
-type WSMsgDispatcher struct {
-	hub *ws.WebsocketHub
-}
-
-func (m *WSMsgDispatcher) Dispatch(datasite string, msg *syftmsg.Message) bool {
-	return m.hub.SendMessageUser(datasite, msg)
-}
-
-func NewWSMsgDispatcher(hub *ws.WebsocketHub) MessageDispatcher {
-	return &WSMsgDispatcher{hub: hub}
-}
-
-func NewBlobMsgStore(blob *blob.BlobService) RPCMsgStore {
-	return &BlobMsgStore{blob: blob}
-}
-
-// SendService handles the business logic for message sending and polling
-type SendService struct {
-	dispatcher MessageDispatcher
-	store      RPCMsgStore
-	cfg        *Config
-}
-
 // Config holds the service configuration
 type Config struct {
 	DefaultTimeoutMs    int
@@ -85,6 +27,13 @@ type Config struct {
 	MaxBodySize         int64
 	PollIntervalMs      int
 	RequestChkTimeoutMs int
+}
+
+// SendService handles the business logic for message sending and polling
+type SendService struct {
+	dispatcher MessageDispatcher
+	store      RPCMsgStore
+	cfg        *Config
 }
 
 // NewSendService creates a new send service
